requirements.txt
torch
torchvision
transformers
Pillow
numpy
tqdm
matplotlib
